Refactor: Extract measurement views from Session Header
