+++ AirCasting/SessionViews/SessionHeaderView.swift
@@ -29,7 +29,8 @@ struct SessionHeaderView: View {
                     }
                 }
             } else {
-                measurementsAB
+                ABMeasurementsView(session: session,
+                                   thresholds: thresholds)
             }
         }
         .font(Font.moderate(size: 13, weight: .regular))
@@ -74,46 +75,12 @@ private extension SessionHeaderView {
         .foregroundColor(.darkBlue)
     }
     
-    var measurementsTitle: some View {
-        Text("Most recent measurement:")
-    }
-    
-    var measurementsAB: some View {
-        Group {
-            if let measurements = extractLatestMeasurements() {
-                VStack(alignment: .leading, spacing: 5) {
-                    measurementsTitle
-                    HStack {
-                        Group {
-                            singleMeasurement(streamName: "PM1", value: measurements.pm1)
-                            singleMeasurement(streamName: "PM2", value: measurements.pm25)
-                            singleMeasurement(streamName: "PM10", value: measurements.pm10)
-                            singleMeasurement(streamName: "F", value: measurements.f)
-                            singleMeasurement(streamName: "RH", value: measurements.h)
-                        }
-                        .frame(maxWidth: .infinity)
-                    }
-                }
-            } else {
-                VStack(alignment: .leading, spacing: 3) {
-                    Text("Your AirBeam is gathering data.")
-                        .font(Font.moderate(size: 14))
-                    Text("Measurements will appear in 3 minutes.")
-                        .font(Font.moderate(size: 12))
-                }
-                .foregroundColor(.darkBlue)
-            }
-        }
-    }
-    
-    func thresholdFor(name: String) -> SensorThreshold? {
-        thresholds.first { $0.sensorName == name }
-    }
-    
     var measurementsMic: some View {
         VStack(alignment: .leading, spacing: 5) {
-            measurementsTitle
-            singleMeasurement(streamName: "db", value: lastMicMeasurement())
+            Text("Most recent measurement:")
+            SingleMeasurementView(streamName: "db",
+                                  value: lastMicMeasurement(),
+                                  thresholds: thresholds)
         }
     }
     
@@ -125,46 +92,7 @@ private extension SessionHeaderView {
                 .foregroundColor(.blue)
         })
     }
-    
-    func singleMeasurement(streamName: String, value: Double) -> some View {
-        VStack(spacing: 3) {
-            Text(streamName)
-                .font(Font.system(size: 13))
-            HStack(spacing: 3){
-                MeasurementDotView(value: value,
-                                   thresholds: thresholdFor(name: streamName))
-                Text("\(Int(value))")
-                    .font(Font.moderate(size: 14, weight: .regular))
-            }
-        }
-    }
-    
-    struct LatestMeasurements {
-        let pm1: Double
-        let pm25: Double
-        let pm10: Double
-        let f: Double
-        let h: Double
-    }
-    func extractLatestMeasurements() -> LatestMeasurements? {
-        let pm1Value = session.pm1Stream?.latestValue ?? 0
-        let pm25Value = session.pm2Stream?.latestValue ?? 0
-        let pm10Value = session.pm10Stream?.latestValue ?? 0
-        let fValue = session.FStream?.latestValue ?? 0
-        let hValue = session.HStream?.latestValue ?? 0
         
-        #warning("TODO: change logic here (session status)")
-        if pm1Value != 0 || pm25Value != 0 || pm10Value != 0 || fValue != 0 || hValue != 0 {
-            return LatestMeasurements(pm1: pm1Value,
-                                      pm25: pm25Value,
-                                      pm10: pm10Value,
-                                      f: fValue,
-                                      h: hValue)
-        } else  {
-            return nil
-        }
-    }
-    
     func lastMicMeasurement() -> Double {
         return session.dbStream?.latestValue ?? 0
     }
